refactor: create SpecOps-agnostic `evmdebug` package, moving relevant code out of `runopts` (excluding UI)
