+++ evmdebug/evmdebug.go
@@ -1,4 +1,7 @@
-package runopts
+// Package evmdebug provides debugging mechanisms for EVM contracts,
+// intercepting opcode-level execution and allowing for inspection of data such
+// as the VM's stack and memory.
+package evmdebug
 
 import (
 	"sync"
@@ -49,8 +52,9 @@ type (
 	done        struct{}
 )
 
-// A Debugger is an Option that intercepts opcode execution to allow inspection
-// of the stack, memory, etc.
+// A Debugger intercepts EVM opcode execution to allow inspection of the stack,
+// memory, etc. The value returned by its Tracer() method should be placed
+// inside a vm.Config before execution commences.
 type Debugger struct {
 	d *debugger
 
@@ -63,11 +67,11 @@ type Debugger struct {
 	done    <-chan done
 }
 
-// Apply adds a VMConfig.Tracer to the Configuration, intercepting execution of
-// every opcode.
-func (d *Debugger) Apply(c *Configuration) error {
-	c.VMConfig.Tracer = d.d
-	return nil
+// Tracer returns an EVMLogger that enables debugging, compatible with geth.
+//
+// TODO: add an example demonstrating how to access vm.Config.
+func (d *Debugger) Tracer() vm.EVMLogger {
+	return d.d
 }
 
 // Wait blocks until the bytecode is ready for execution, but the first opcode
